Create GameVersionNumber
