+++ HMCL/src/main/java/org/jackhuang/hmcl/upgrade/hmcl/UpdateChecker.java
@@ -32,7 +32,7 @@ import static org.jackhuang.hmcl.util.Lang.mapOf;
 import static org.jackhuang.hmcl.util.Lang.thread;
 import static org.jackhuang.hmcl.util.Logging.LOG;
 import static org.jackhuang.hmcl.util.Pair.pair;
-import static org.jackhuang.hmcl.util.versioning.VersionNumber.asVersion;
+import static org.jackhuang.hmcl.util.versioning.DefaultVersionNumber.asVersion;
 
 public final class UpdateChecker {
     private UpdateChecker() {}
