+++ HMCLCore/src/main/java/org/jackhuang/hmcl/util/versioning/GameVersionNumber.java
@@ -1,0 +1,23 @@
+package org.jackhuang.hmcl.util.versioning;
+
+public abstract class GameVersionNumber {
+
+    GameVersionNumber() {
+    }
+
+
+    abstract Type getType();
+
+    enum Type {
+        PRE_CLASSIC,
+
+    }
+
+
+    static final class PreClassic extends GameVersionNumber {
+        @Override
+        Type getType() {
+            return Type.PRE_CLASSIC;
+        }
+    }
+}
